Add exploit_tests.cpp
This new test is a catch all to test for any exploit type testing.

-Also move the p2p message exploit tests from thinblock_tests.cpp to
this test file.

-Make receive queues and ip addresses reuseable in exploit_tests.cpp

-Add tests for block consistency checking

Consistency checks for blocks are handled fully in checkblock_tests.cpp.

-Also cleaned up the formatting and added some comments to block handling
in main.cpp

-Add transaction consistency checks

These are fully covered in transaction_tests.cpp and
AcceptToMempool()
