+++ src/test/thinblock_tests.cpp
@@ -25,13 +25,6 @@
 
 extern void BuildSeededBloomFilter(CBloomFilter& memPoolFilter, std::vector<uint256>& vOrphanHashes, uint256 hash);
 
-CService ipaddress(uint32_t i)
-{
-    struct in_addr s;
-    s.s_addr = i;
-    return CService(CNetAddr(s), Params().GetDefaultPort());
-}
-
 CBlock TestBlock() { //Thanks dagurval :)
     // Block taken from bloom_tests.cpp merkle_block_1
     // Random real block (0000000000013b8ab2cd513b0261a14096412195a72a0c4827d229dcc7e0f7af)
@@ -41,6 +34,7 @@ CBlock TestBlock() { //Thanks dagurval :)
     stream >> block;
     return block;
 };
+
 
 BOOST_FIXTURE_TEST_SUITE(thinblock_tests, TestingSetup)
 
@@ -125,232 +119,6 @@ BOOST_AUTO_TEST_CASE(thinblock_test) {
     filter1.clear();
     CXThinBlock xthinblock7(block, &filter1);
     BOOST_CHECK(xthinblock7.collision);
-
-    /* Message Consistency checking for p2p thinblock messages.  If the messages are not formatted correctly then a
-       a ban will result.
-     */
-    fImporting = false;
-    fReindex = false;
-
-    /** XTHINBLOCK message consistency checks */
-
-    // testing empty missingtx vector
-    CAddress addr1(ipaddress(0xa0b0c001));
-    CNode dummyNode1(INVALID_SOCKET, addr1, "", true);
-    dummyNode1.nVersion = 1;
-
-    CNode::ClearBanned();
-    CDataStream vRecv1(SER_NETWORK, PROTOCOL_VERSION);
-    CXThinBlock xthin = xthinblock;
-    xthin.vMissingTx.clear(); // empty the missingtx vector. This should cause an error.
-    vRecv1 << xthin;
-    ProcessMessage(&dummyNode1, NetMsgType::XTHINBLOCK, vRecv1, GetTime());
-    SendMessages(&dummyNode1);
-    BOOST_CHECK(xthin.vMissingTx.size() == 0);
-    BOOST_CHECK(CNode::IsBanned(addr1));
-    
-    // test invalid or missing coinbase
-    CAddress addr1a(ipaddress(0xa0b0c01a));
-    CNode dummyNode1a(INVALID_SOCKET, addr1a, "", true);
-    dummyNode1a.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv1.clear();
-    xthin = xthinblock;
-    xthin.vMissingTx[0] = xthin.vMissingTx[1]; // delete the coinbase. This should cause an error.
-    vRecv1 << xthin;
-    ProcessMessage(&dummyNode1a, NetMsgType::XTHINBLOCK, vRecv1, GetTime());
-    SendMessages(&dummyNode1a);
-    BOOST_CHECK(!xthin.vMissingTx[0].IsCoinBase());
-    BOOST_CHECK(CNode::IsBanned(addr1a));
-
-    // test invalid block header
-    CAddress addr1b(ipaddress(0xa0b0c01b));
-    CNode dummyNode1b(INVALID_SOCKET, addr1b, "", true);
-    dummyNode1b.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv1.clear();
-    xthin = xthinblock;
-    xthin.header.nBits = 1; // create invalid block header
-    vRecv1 << xthin;
-    ProcessMessage(&dummyNode1b, NetMsgType::XTHINBLOCK, vRecv1, GetTime());
-    SendMessages(&dummyNode1b);
-    CValidationState state;
-    BOOST_CHECK(!CheckBlockHeader(xthin.header, state, true));
-    BOOST_CHECK(CNode::IsBanned(addr1b));
-
-    /** THINBLOCK message consistency  checks */
-
-    // test empty missingtx vector
-    CAddress addr2(ipaddress(0xa0b0c02a));
-    CNode dummyNode2(INVALID_SOCKET, addr2, "", true);
-    dummyNode2.nVersion = 1;
-
-    CNode::ClearBanned();
-    CDataStream vRecv2(SER_NETWORK, PROTOCOL_VERSION);
-    CThinBlock thin = thinblock;
-    thin.vMissingTx.clear(); // empty the missingtx vector. This should cause an error.
-    vRecv2 << thin;
-    ProcessMessage(&dummyNode2, NetMsgType::THINBLOCK, vRecv2, GetTime());
-    SendMessages(&dummyNode2);
-    BOOST_CHECK(thin.vMissingTx.size() == 0);
-    BOOST_CHECK(CNode::IsBanned(addr2));
-
-    // test invalid or missing coinbase
-    CAddress addr2b(ipaddress(0xa0b0c02b));
-    CNode dummyNode2b(INVALID_SOCKET, addr2b, "", true);
-    dummyNode2b.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv2.clear();
-    thin = thinblock;
-    thin.vMissingTx[0] = thin.vMissingTx[1]; // delete the coinbase. This should cause an error.
-    vRecv2 << thin;
-    ProcessMessage(&dummyNode2b, NetMsgType::THINBLOCK, vRecv2, GetTime());
-    SendMessages(&dummyNode2b);
-    BOOST_CHECK(!thin.vMissingTx[0].IsCoinBase());
-    BOOST_CHECK(CNode::IsBanned(addr2b));
-
-    // create invalid block header
-    CAddress addr2c(ipaddress(0xa0b0c02c));
-    CNode dummyNode2c(INVALID_SOCKET, addr2c, "", true);
-    dummyNode2c.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv2.clear();
-    thin = thinblock;
-    thin.header.nBits = 1;
-    vRecv2 << thin;
-    ProcessMessage(&dummyNode2c, NetMsgType::THINBLOCK, vRecv2, GetTime());
-    SendMessages(&dummyNode2c);
-    BOOST_CHECK(!CheckBlockHeader(thin.header, state, true));
-    BOOST_CHECK(CNode::IsBanned(addr2c));
-
-
-    /** XBLOCKTX message consistency checks */
-
-    // test null hash
-    CBlock block3 = TestBlock();
-    CAddress addr3(ipaddress(0xa0b0c003));
-    CNode dummyNode3(INVALID_SOCKET, addr3, "", true);
-    dummyNode3.nVersion = 1;
-
-    CNode::ClearBanned();
-    uint256 nullhash;
-    nullhash.SetNull();
-    CDataStream vRecv3(SER_NETWORK, PROTOCOL_VERSION);
-    CXThinBlockTx xblocktx(nullhash, block3.vtx);
-    vRecv3 << xblocktx;
-    ProcessMessage(&dummyNode3, NetMsgType::XBLOCKTX, vRecv3, GetTime());
-    SendMessages(&dummyNode3);
-    BOOST_CHECK(nullhash.IsNull());
-    BOOST_CHECK(CNode::IsBanned(addr3));
-
-    // test no txns in xblocktx
-    CAddress addr3a(ipaddress(0xa0b0c03a));
-    CNode dummyNode3a(INVALID_SOCKET, addr3a, "", true);
-    dummyNode3a.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv3.clear();
-    std::vector<CTransaction> vTxEmpty;
-    CXThinBlockTx xblocktx2(block3.GetHash(), vTxEmpty);
-    vRecv3 << xblocktx2;
-    ProcessMessage(&dummyNode3a, NetMsgType::XBLOCKTX, vRecv3, GetTime());
-    SendMessages(&dummyNode3a);
-    BOOST_CHECK(vTxEmpty.size() == 0);
-    BOOST_CHECK(CNode::IsBanned(addr3a));
-
-    // test txns hashes in xblocktx not matching num in pfrom->xThinBlockHashes
-    CAddress addr3b(ipaddress(0xa0b0c03b));
-    CNode dummyNode3b(INVALID_SOCKET, addr3b, "", true);
-    dummyNode3b.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv3.clear();
-    CXThinBlockTx xblocktx3(block3.GetHash(), block3.vtx);
-    vRecv3 << xblocktx3;
-    dummyNode3b.xThinBlockHashes.push_back(1); // add one hash to the vector which will cause a mismatch
-    ProcessMessage(&dummyNode3b, NetMsgType::XBLOCKTX, vRecv3, GetTime());
-    SendMessages(&dummyNode3b);
-    BOOST_CHECK(dummyNode3b.xThinBlockHashes.size() != dummyNode3b.thinBlock.vtx.size());
-    BOOST_CHECK(CNode::IsBanned(addr3b));
-
-
-    /** GET_XBLOCKTX message consistency checks */
-
-    // test null hash
-    CAddress addr4a(ipaddress(0xa0b0c04a));
-    CNode dummyNode4a(INVALID_SOCKET, addr4a, "", true);
-    dummyNode4a.nVersion = 1;
-
-    CNode::ClearBanned();
-    nullhash.SetNull();
-    CDataStream vRecv4(SER_NETWORK, PROTOCOL_VERSION);
-    std::set<uint64_t> setHashesToRequest;
-    setHashesToRequest.insert(1); // add a hash so that we are not empty
-    CXRequestThinBlockTx get_xblocktx(nullhash, setHashesToRequest);
-    vRecv4 << get_xblocktx;
-    ProcessMessage(&dummyNode4a, NetMsgType::GET_XBLOCKTX, vRecv4, GetTime());
-    SendMessages(&dummyNode4a);
-    BOOST_CHECK(nullhash.IsNull());
-    BOOST_CHECK(CNode::IsBanned(addr4a));
-
-    // test empty setHashesToRequest
-    CAddress addr4b(ipaddress(0xa0b0c04b));
-    CNode dummyNode4b(INVALID_SOCKET, addr4b, "", true);
-    dummyNode4b.nVersion = 1;
-
-    CNode::ClearBanned();
-    vRecv4.clear();
-    setHashesToRequest.clear(); // clear the set
-    CXRequestThinBlockTx get_xblocktx2(block3.GetHash(), setHashesToRequest);
-    vRecv4 << get_xblocktx2;
-    ProcessMessage(&dummyNode4b, NetMsgType::GET_XBLOCKTX, vRecv4, GetTime());
-    SendMessages(&dummyNode4b);
-    BOOST_CHECK(setHashesToRequest.empty());
-    BOOST_CHECK(CNode::IsBanned(addr4b));
-
-
-    /** GET_XTHIN message consistency checks */
-
-    // test get_xthin with null hash
-    CAddress addr5a(ipaddress(0xa0b0c05a));
-    CNode dummyNode5a(INVALID_SOCKET, addr5a, "", true);
-    dummyNode5a.nVersion = 1;
-
-
-    CNode::ClearBanned();
-    nullhash.SetNull();
-    CInv inv(MSG_XTHINBLOCK, nullhash);
-    CDataStream vRecv5(SER_NETWORK, PROTOCOL_VERSION);
-    CBloomFilter filterMemPool;
-    BuildSeededBloomFilter(filterMemPool, vOrphanHashes, inv.hash, true);
-    vRecv5 << inv;
-    vRecv5 << filterMemPool;
-    ProcessMessage(&dummyNode5a, NetMsgType::GET_XTHIN, vRecv5, GetTime());
-    SendMessages(&dummyNode5a);
-    BOOST_CHECK(nullhash.IsNull());
-    BOOST_CHECK(CNode::IsBanned(addr5a));
-
-    // test get_xthin with invalid message type
-    CAddress addr5b(ipaddress(0xa0b0c05b));
-    CNode dummyNode5b(INVALID_SOCKET, addr5b, "", true);
-    dummyNode5b.nVersion = 1;
-
-
-    CNode::ClearBanned();
-    vRecv5.clear();
-    CInv inv2(15, block3.GetHash()); // invalid type
-    CBloomFilter filterMemPool2;
-    BuildSeededBloomFilter(filterMemPool2, vOrphanHashes, inv2.hash, true);
-    vRecv5 << inv2;
-    vRecv5 << filterMemPool2;
-    ProcessMessage(&dummyNode5b, NetMsgType::GET_XTHIN, vRecv5, GetTime());
-    SendMessages(&dummyNode5b);
-    BOOST_CHECK(inv2.type != MSG_THINBLOCK && inv2.type != MSG_XTHINBLOCK);
-    BOOST_CHECK(CNode::IsBanned(addr5b));
 }
 
 BOOST_AUTO_TEST_SUITE_END()
